+++ server/testingHelpers.js
@@ -155,12 +155,26 @@ async function firmwareAlert(chai, server, coreID, sensorEvent, apiKey) {
 }
 
 function mockResponse(sandbox) {
-  // From https://codewithhugo.com/express-request-response-mocking/
   const res = {}
+
   res.writeHead = sandbox.stub().returns(res)
-  res.json = sandbox.stub().returns(res)
   res.status = sandbox.stub().returns(res)
-  res.send = sandbox.stub().returns(res)
+
+  // for more rigorous testing, res.body will be
+  // set to the arguments to res.json and res.send
+  res.body = {}
+
+  res.json = sandbox.stub().callsFake(json => {
+    res.body = json
+
+    return res
+  })
+
+  res.send = sandbox.stub().callsFake(data => {
+    res.body = data
+
+    return res
+  })
 
   return res
 }